chore: Split Startup page
Startup page has been split into two extra components, "Welcome" which displays the logo and Welcome text and "SignIn" which handles displaying components for signing in as well as the logic for it.
Closes #24
